+++ subgrade-impl/src/main/java/com/dwarfeng/subgrade/impl/dao/HibernateBatchRelationDao.java
@@ -18,7 +18,9 @@ import java.util.stream.Collectors;
 
 /**
  * 使用 Hibernate 基于多对多关系实现的 BatchRelationDao。
- * <p>该类只提供最基本的方法实现，没有添加任何事务，请通过代理的方式在代理类中添加事务。</p>
+ *
+ * <p>
+ * 该类只提供最基本的方法实现，没有添加任何事务，请通过代理的方式在代理类中添加事务。
  *
  * @author DwArFeng
  * @since 0.2.4-beta
@@ -45,15 +47,25 @@ public class HibernateBatchRelationDao<
         JOIN_BY_CHILD
     }
 
+    @Nonnull
     private HibernateTemplate template;
+    @Nonnull
     private BeanTransformer<PK, PPK> pkTransformer;
+    @Nonnull
     private BeanTransformer<CK, PCK> ckTransformer;
+    @Nonnull
     private BeanTransformer<PE, PPE> peTransformer;
+    @Nonnull
     private BeanTransformer<CE, PCE> ceTransformer;
+    @Nonnull
     private Class<PPE> classPPE;
+    @Nonnull
     private Class<PCE> classPCE;
+    @Nonnull
     private String parentProperty;
+    @Nonnull
     private String childProperty;
+    @Nonnull
     private JoinType joinType;
 
     public HibernateBatchRelationDao(
@@ -66,7 +78,8 @@ public class HibernateBatchRelationDao<
             @Nonnull Class<PCE> classPCE,
             @Nonnull String parentProperty,
             @Nonnull String childProperty,
-            @Nonnull JoinType joinType) {
+            @Nonnull JoinType joinType
+    ) {
         this.template = template;
         this.pkTransformer = pkTransformer;
         this.ckTransformer = ckTransformer;
@@ -263,6 +276,7 @@ public class HibernateBatchRelationDao<
         }
     }
 
+    @Nonnull
     public HibernateTemplate getTemplate() {
         return template;
     }
@@ -271,6 +285,7 @@ public class HibernateBatchRelationDao<
         this.template = template;
     }
 
+    @Nonnull
     public BeanTransformer<PK, PPK> getPkTransformer() {
         return pkTransformer;
     }
@@ -279,6 +294,7 @@ public class HibernateBatchRelationDao<
         this.pkTransformer = pkTransformer;
     }
 
+    @Nonnull
     public BeanTransformer<CK, PCK> getCkTransformer() {
         return ckTransformer;
     }
@@ -287,6 +303,7 @@ public class HibernateBatchRelationDao<
         this.ckTransformer = ckTransformer;
     }
 
+    @Nonnull
     public BeanTransformer<PE, PPE> getPeTransformer() {
         return peTransformer;
     }
@@ -295,6 +312,7 @@ public class HibernateBatchRelationDao<
         this.peTransformer = peTransformer;
     }
 
+    @Nonnull
     public BeanTransformer<CE, PCE> getCeTransformer() {
         return ceTransformer;
     }
@@ -303,6 +321,7 @@ public class HibernateBatchRelationDao<
         this.ceTransformer = ceTransformer;
     }
 
+    @Nonnull
     public Class<PPE> getClassPPE() {
         return classPPE;
     }
@@ -311,6 +330,7 @@ public class HibernateBatchRelationDao<
         this.classPPE = classPPE;
     }
 
+    @Nonnull
     public Class<PCE> getClassPCE() {
         return classPCE;
     }
@@ -319,6 +339,7 @@ public class HibernateBatchRelationDao<
         this.classPCE = classPCE;
     }
 
+    @Nonnull
     public String getParentProperty() {
         return parentProperty;
     }
@@ -327,6 +348,7 @@ public class HibernateBatchRelationDao<
         this.parentProperty = parentProperty;
     }
 
+    @Nonnull
     public String getChildProperty() {
         return childProperty;
     }
@@ -335,6 +357,7 @@ public class HibernateBatchRelationDao<
         this.childProperty = childProperty;
     }
 
+    @Nonnull
     public JoinType getJoinType() {
         return joinType;
     }
@@ -345,7 +368,7 @@ public class HibernateBatchRelationDao<
 
     @Override
     public String toString() {
-        return "HibernateMtmBatchRelationDao{" +
+        return "HibernateBatchRelationDao{" +
                 "template=" + template +
                 ", pkTransformer=" + pkTransformer +
                 ", ckTransformer=" + ckTransformer +