+++ subgrade-impl/src/main/java/com/dwarfeng/subgrade/impl/dao/RedisBaseDao.java
@@ -13,22 +13,32 @@ import javax.annotation.Nonnull;
 
 /**
  * 通过 Redis 实现的基础数据访问层。
- * <p>由于数据存放在内存中，请不要使用该数据访问层存储大量的数据。</p>
- * <p>该类只提供最基本的方法实现，没有添加任何事务，请通过代理的方式在代理类中添加事务。</p>
+ *
+ * <p>
+ * 由于数据存放在内存中，请不要使用该数据访问层存储大量的数据。
+ *
+ * <p>
+ * 该类只提供最基本的方法实现，没有添加任何事务，请通过代理的方式在代理类中添加事务。
  *
  * @author DwArFeng
  * @since 0.0.3-beta
  */
 public class RedisBaseDao<K extends Key, E extends Entity<K>, JE extends Bean> implements BaseDao<K, E> {
 
+    @Nonnull
     private RedisTemplate<String, JE> template;
+    @Nonnull
     private StringKeyFormatter<K> formatter;
+    @Nonnull
     private BeanTransformer<E, JE> transformer;
+    @Nonnull
     private String dbKey;
 
     public RedisBaseDao(
-            @Nonnull RedisTemplate<String, JE> template, @Nonnull StringKeyFormatter<K> formatter,
-            @Nonnull BeanTransformer<E, JE> transformer, @Nonnull String dbKey
+            @Nonnull RedisTemplate<String, JE> template,
+            @Nonnull StringKeyFormatter<K> formatter,
+            @Nonnull BeanTransformer<E, JE> transformer,
+            @Nonnull String dbKey
     ) {
         this.template = template;
         this.formatter = formatter;
@@ -101,6 +111,7 @@ public class RedisBaseDao<K extends Key, E extends Entity<K>, JE extends Bean> i
         }
     }
 
+    @Nonnull
     public RedisTemplate<String, JE> getTemplate() {
         return template;
     }
@@ -109,6 +120,7 @@ public class RedisBaseDao<K extends Key, E extends Entity<K>, JE extends Bean> i
         this.template = template;
     }
 
+    @Nonnull
     public StringKeyFormatter<K> getFormatter() {
         return formatter;
     }
@@ -117,6 +129,7 @@ public class RedisBaseDao<K extends Key, E extends Entity<K>, JE extends Bean> i
         this.formatter = formatter;
     }
 
+    @Nonnull
     public BeanTransformer<E, JE> getTransformer() {
         return transformer;
     }
@@ -125,11 +138,22 @@ public class RedisBaseDao<K extends Key, E extends Entity<K>, JE extends Bean> i
         this.transformer = transformer;
     }
 
+    @Nonnull
     public String getDbKey() {
         return dbKey;
     }
 
     public void setDbKey(@Nonnull String dbKey) {
         this.dbKey = dbKey;
+    }
+
+    @Override
+    public String toString() {
+        return "RedisBaseDao{" +
+                "template=" + template +
+                ", formatter=" + formatter +
+                ", transformer=" + transformer +
+                ", dbKey='" + dbKey + '\'' +
+                '}';
     }
 }