+++ subgrade-impl/src/main/java/com/dwarfeng/subgrade/impl/dao/RedisEntireLookupDao.java
@@ -16,17 +16,25 @@ import java.util.stream.Collectors;
 
 /**
  * 通过 Redis 实现的 EntireLookupDao。
- * <p>由于数据存放在内存中，请不要使用该数据访问层存储大量的数据。</p>
- * <p>该类只提供最基本的方法实现，没有添加任何事务，请通过代理的方式在代理类中添加事务。</p>
+ *
+ * <p>
+ * 由于数据存放在内存中，请不要使用该数据访问层存储大量的数据。
+ *
+ * <p>
+ * 该类只提供最基本的方法实现，没有添加任何事务，请通过代理的方式在代理类中添加事务。
  *
  * @author DwArFeng
  * @since 0.0.3-beta
  */
 public class RedisEntireLookupDao<K extends Key, E extends Entity<K>, JE extends Bean> implements EntireLookupDao<E> {
 
+    @Nonnull
     private RedisTemplate<String, JE> template;
+    @Nonnull
     private StringKeyFormatter<K> formatter;
+    @Nonnull
     private BeanTransformer<E, JE> transformer;
+    @Nonnull
     private String dbKey;
 
     public RedisEntireLookupDao(
@@ -91,6 +99,7 @@ public class RedisEntireLookupDao<K extends Key, E extends Entity<K>, JE extends
         }
     }
 
+    @Nonnull
     public RedisTemplate<String, JE> getTemplate() {
         return template;
     }
@@ -99,6 +108,7 @@ public class RedisEntireLookupDao<K extends Key, E extends Entity<K>, JE extends
         this.template = template;
     }
 
+    @Nonnull
     public StringKeyFormatter<K> getFormatter() {
         return formatter;
     }
@@ -107,6 +117,7 @@ public class RedisEntireLookupDao<K extends Key, E extends Entity<K>, JE extends
         this.formatter = formatter;
     }
 
+    @Nonnull
     public BeanTransformer<E, JE> getTransformer() {
         return transformer;
     }
@@ -115,11 +126,22 @@ public class RedisEntireLookupDao<K extends Key, E extends Entity<K>, JE extends
         this.transformer = transformer;
     }
 
+    @Nonnull
     public String getDbKey() {
         return dbKey;
     }
 
     public void setDbKey(@Nonnull String dbKey) {
         this.dbKey = dbKey;
+    }
+
+    @Override
+    public String toString() {
+        return "RedisEntireLookupDao{" +
+                "template=" + template +
+                ", formatter=" + formatter +
+                ", transformer=" + transformer +
+                ", dbKey='" + dbKey + '\'' +
+                '}';
     }
 }